Add basic mark and sweep GC
Removes dependence on Boehm GC, which is nice. Performance isn't a huge
priority anyways.

As set up now, the collector runs after 16 kiB has been allocated since
the GC was last run. It doesn't consider the whole C stack in order to
find references to any cheax object. If the GC is run while an object
remains only in use through the C API, it will be destroyed. That is,
unless you've used cheax_ref() to add an extra strong reference to the
object. Make sure to use cheax_unref() once you're done with it.

The GC is only run as part of cheax_eval().