Add revamped symbol lookup system, improve macros
The previous system had been in place since 2016, and even back then I
knew it wasn't ideal. Lookups were O(# of symbols). The new system
uses binary trees, so lookups are O(log(# of symbols)).

As an added benefit, function call logic in eval.c and api.c is much
easier to understand.

The new system introduces environments (struct chx_env, type CHEAX_ENV),
which form a rather special kind of chx_value, namely the only truly
mutable kind. Environments associate symbols with values, and can be
"called" like a function to enter them. A new environment is entered
automatically when a function is called (after arguments are evaluated),
and left when said function returns. But here's the kicker: environments
are just values, and can thus be stored. One can use the new (env)
builtin to obtain the current environment, and store it somehwere or
return it. Thus, there is now a neat way to make records:

  (defun vec2 (_x _y)
    (var x _x)
    (var y _y)
    (defun add (other)
      (vec2 (+ x (other x)) (+ y (other y))))
    (defun str ()
      (format "(vec2 {} {})" x y))

    ; now return the environment in which the above was declared
    (env))

  (def vec_a (vec2 100 200))
  (def vec_b (vec2 200 300))
  (def vec_c (vec_a (add vec_b)))
  (vec_c (str))                   ; => "(vec2 300 500)"

Refactored a lot of code in eval.c (also code that had been there
since 2016), and make sure macros behave like they should in any lisp.
Previously, besides macros not evaluating their argument list when
called (which is the correct behaviour), the expressions in their bodies
were left unevaluated. Rather, a substitution was performed (rather like
with backquotes, but with commas added automatically). Then the result
was evaluated in the caller context. This is what my C-brain thought
macros were. Now, body expressions in macros are evaluated, and indeed
the return value of a macro is re-evaluated in the caller context.
For example:

  (defmacro foo (x) 'x)
  (var y 0)

  (foo y) => y                         ; old behaviour: calling macro
                                       ; would lead to substitution of
				       ; 'x with 'y, which was then
				       ; evaluated in caller context.
  (foo y) => ENOSYM no such symbol `x' ; new behaviour: macro simply
                                       ; returns x, no substitution is
				       ; performed if backquotes aren't
				       ; used.

Other changes:
- Add CHEAX_TYPE_MASK, indicating which bits of the chx_value::type are
  actually used to name types.
- Add bit flags that can be set in chx_value::type for internal use:
  - NO_GC_BIT:   value was not allocated by GC
  - FIN_BIT:     value has registered finaliser function
  - BIF_ENV_BIT: value (which is a chx_env) is bifurcated environment,
                 rather than a regular one.
- Add set_type() function, which is mindful of CHEAX_TYPE_MASK.
- Remove named enum chx_varflags, replace with int and anyonymous enum.