Add new symbol system, (defsym) builtin
This has been on my mind for quite a while now. The whole `struct
variable' system dates from 2016, and I think it's run its course. It
was nice as a very quick solution to the problem of synchronising
variables from C into cheax.

This new system is a bit simpler, in some ways, as well as more
versatile. Previously, struct variable would store the symbol name, a
field containing some flags, and then a union which would store either
just a regular chx_value, or a pointer to something in C with which it
was synchronized. In the new system, a symbol is in essence just a
struct with a getter and a setter, and a pointer to some user-definable
data. This means that the cheax_sync*() API can remain as is, but one
now also has the option to define custom symbols with custom getters and
setters.

This ability to have symbols with custom getters and setters is also
exposed to the cheax environment itself, through the (defsym) builtin.
An example:

  (var temp-c 20.0)    ; temperature in celsius
  (defsym temp-f       ; temperature in fahrenheit
    (def factor 1.8)   ; just a useless example to highlight that
                       ; defsym enters a new scope in which custom
                       ; symbols and functions can be declared
    (defget (+ 32.0 (* temp-c factor)))
    (getset (set temp-c (/ (- value 32.0) factor))))
                            ; ^ a special id containing the arg to (set)

  temp-f               ; => 68.0
  (set temp-f 0.0)
  temp-c               ; => -17.777778