Free func call environments early; lower gc load
This method is rather dangerous, but should be worth it. Most function
call environments are created and swiftly turned into garbage. The only
way that doesn't happen is if the value of the environment escape
(through the (env) builtin). This optimisation catches that situation,
and gcol_free()'s the environment otherwise.