Fix unpack() not resetting storage behaviour
I.e.

  struct chx_value num;
  struct chx_list *list;
  unpack("[id]c", &num, &list);

Would fail, since "[id]" requires output stored as chx_value (instead of
as chx_int or chx_double), but that behaviour wouldn't reset for the
second argument. So the list would also be stored as a chx_value,
instead of as a chx_list *. Oops! Guess it hasn't been a problem so
far...