Re-implement CHEAX_EXT_FUNC
Obviously this should be allowed

  (map floor '(1.2 1.3))

While this should not:

  (map cond '(() ()))

By making (floor) of type CHEAX_EXT_FUNC and (cond) of type
CHEAX_SPECIAL_FORM (or, more accurately, CHEAX_SPECIAL_TAIL_FORM), one
can disallow the second with a type-assert in map. This is indeed what
map does.

In the end, special forms should be handled by some kind of macro
expansion function, and not be exposed at runtime at all. But this will
do for now.