Replace red-black trees with hash tables
Red-black trees were not the "correct" data structure for an associative
structure in this programme, since we never need things to remain
ordered.

This change speeds up things by about 11% on my machine.

The goal was to make a dynamic hash table that is simple, allows for a
quick if-not-present-then-insert idiom (i.e. does not need to do a
look-up twice in such a case), and that is very cheap to create. The
latter requirement is because each call to cheax_push_env() initalizes
a hash table, and indeed this function gets called a lot. Furthermore,
the hash tables created by cheax_push_env() likely remain very small
(will perhaps even remain empty).

Suggestions for improvement: clean-up sym.c to that htab_get() is only
called once when a symbol is defined.